skip float64 tests

+++ heat/core/tests/test_io.py
@@ -811,17 +811,19 @@ class TestIO(TestCase):
             float_array = np.concatenate(crea_array, 1)
         ht.MPI_WORLD.Barrier()
 
-        load_array = ht.load_npy_from_path(
-            os.path.join(os.getcwd(), "heat/datasets"), dtype=ht.float64, split=1
-        )
-        load_array_npy = load_array.numpy()
-        self.assertIsInstance(load_array, ht.DNDarray)
-        self.assertEqual(load_array.dtype, ht.float64)
-        if ht.MPI_WORLD.rank == 0:
-            self.assertTrue((load_array_npy == float_array).all)
-            for file in os.listdir(os.path.join(os.getcwd(), "heat/datasets")):
-                if fnmatch.fnmatch(file, "*.npy"):
-                    os.remove(os.path.join(os.getcwd(), "heat/datasets", file))
+        if not self.is_mps:
+            # float64 not supported in MPS
+            load_array = ht.load_npy_from_path(
+                os.path.join(os.getcwd(), "heat/datasets"), dtype=ht.float64, split=1
+            )
+            load_array_npy = load_array.numpy()
+            self.assertIsInstance(load_array, ht.DNDarray)
+            self.assertEqual(load_array.dtype, ht.float64)
+            if ht.MPI_WORLD.rank == 0:
+                self.assertTrue((load_array_npy == float_array).all)
+                for file in os.listdir(os.path.join(os.getcwd(), "heat/datasets")):
+                    if fnmatch.fnmatch(file, "*.npy"):
+                        os.remove(os.path.join(os.getcwd(), "heat/datasets", file))
 
     def test_load_npy_exception(self):
         with self.assertRaises(TypeError):
